requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0
sqlparse==0.5.2
tzdata==2024.2
